fix: should not generate css runtime if not enabling css
