+++ crates/rspack_plugin_css/Cargo.toml
@@ -19,6 +19,7 @@ rspack_core = { path = "../rspack_core" }
 rspack_error = { path = "../rspack_error" }
 rspack_hash = { path = "../rspack_hash" }
 rspack_identifier = { path = "../rspack_identifier" }
+rspack_plugin_runtime = { path = "../rspack_plugin_runtime" }
 rustc-hash = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
