+++ crates/rspack_plugin_css/src/lib.rs
@@ -4,8 +4,9 @@
 pub mod dependency;
 mod parser_and_generator;
 pub mod plugin;
+pub mod runtime;
 pub mod swc_css_compiler;
 mod utils;
 pub mod visitors;
 
-pub use plugin::CssPlugin;
+pub use plugin::{CssConfig, CssPlugin};
