+++ crates/rspack_plugin_css/src/plugin/impl_plugin_for_css_plugin.rs
@@ -11,13 +11,16 @@ use rspack_core::{
   SourceType,
 };
 use rspack_core::{
-  Compilation, CompilationArgs, CompilationParams, CompilerOptions, DependencyType,
-  LibIdentOptions, PluginCompilationHookOutput, PublicPath,
+  ChunkLoading, ChunkLoadingType, Compilation, CompilationArgs, CompilationParams, CompilerOptions,
+  DependencyType, LibIdentOptions, PluginCompilationHookOutput, PluginContext,
+  PluginRuntimeRequirementsInTreeOutput, PublicPath, RuntimeGlobals, RuntimeRequirementsInTreeArgs,
 };
 use rspack_error::{IntoTWithDiagnosticArray, Result};
 use rspack_hash::RspackHash;
+use rspack_plugin_runtime::is_enabled_for_chunk;
 
 use crate::parser_and_generator::CssParserAndGenerator;
+use crate::runtime::CssLoadingRuntimeModule;
 use crate::utils::AUTO_PUBLIC_PATH_PLACEHOLDER_REGEX;
 use crate::CssPlugin;
 
@@ -265,5 +268,31 @@ impl Plugin for CssPlugin {
       )]
       .with_empty_diagnostic(),
     )
+  }
+
+  fn runtime_requirements_in_tree(
+    &self,
+    _ctx: PluginContext,
+    args: &mut RuntimeRequirementsInTreeArgs,
+  ) -> PluginRuntimeRequirementsInTreeOutput {
+    let compilation = &mut args.compilation;
+    let chunk = args.chunk;
+    let chunk_loading_value = ChunkLoading::Enable(ChunkLoadingType::Jsonp);
+    let is_enabled_for_chunk = is_enabled_for_chunk(chunk, &chunk_loading_value, compilation);
+    let runtime_requirements = args.runtime_requirements;
+    let runtime_requirements_mut = &mut args.runtime_requirements_mut;
+
+    if (runtime_requirements.contains(RuntimeGlobals::ENSURE_CHUNK_HANDLERS)
+      || runtime_requirements.contains(RuntimeGlobals::HMR_DOWNLOAD_UPDATE_HANDLERS))
+      && is_enabled_for_chunk
+    {
+      runtime_requirements_mut.insert(RuntimeGlobals::PUBLIC_PATH);
+      runtime_requirements_mut.insert(RuntimeGlobals::GET_CHUNK_CSS_FILENAME);
+      runtime_requirements_mut.insert(RuntimeGlobals::HAS_OWN_PROPERTY);
+      runtime_requirements_mut.insert(RuntimeGlobals::MODULE_FACTORIES_ADD_ONLY);
+      compilation.add_runtime_module(chunk, Box::<CssLoadingRuntimeModule>::default());
+    }
+
+    Ok(())
   }
 }
