+++ crates/rspack_plugin_runtime/src/lib.rs
@@ -8,8 +8,6 @@ pub use common_js_chunk_format::CommonJsChunkFormatPlugin;
 mod runtime_plugin;
 use rspack_core::{BoxPlugin, ChunkLoading, ChunkLoadingType, PluginExt};
 pub use runtime_plugin::RuntimePlugin;
-mod css_modules;
-pub use css_modules::CssModulesPlugin;
 mod array_push_callback_chunk_format;
 pub use array_push_callback_chunk_format::ArrayPushCallbackChunkFormatPlugin;
 mod common_js_chunk_loading;
@@ -23,6 +21,9 @@ pub use module_chunk_loading::ModuleChunkLoadingPlugin;
 mod import_scripts_chunk_loading;
 pub use import_scripts_chunk_loading::ImportScriptsChunkLoadingPlugin;
 mod runtime_module;
+pub use runtime_module::{
+  chunk_has_css, is_enabled_for_chunk, render_condition_map, stringify_chunks, BooleanMatcher,
+};
 mod startup_chunk_dependencies;
 pub use startup_chunk_dependencies::StartupChunkDependenciesPlugin;
 mod chunk_prefetch_preload;
@@ -31,7 +32,6 @@ pub use chunk_prefetch_preload::ChunkPrefetchPreloadPlugin;
 pub fn enable_chunk_loading_plugin(loading_type: ChunkLoadingType, plugins: &mut Vec<BoxPlugin>) {
   match loading_type {
     ChunkLoadingType::Jsonp => {
-      plugins.push(CssModulesPlugin.boxed());
       plugins.push(JsonpChunkLoadingPlugin.boxed());
     }
     ChunkLoadingType::Require => {
