+++ crates/rspack_plugin_runtime/src/runtime_module/mod.rs
@@ -9,7 +9,6 @@ mod chunk_preload_trigger;
 mod compat_get_default_export;
 mod create_fake_namespace_object;
 mod create_script_url;
-mod css_loading;
 mod define_property_getters;
 mod ensure_chunk;
 mod export_webpack_require;
@@ -51,7 +50,6 @@ pub use chunk_preload_trigger::ChunkPreloadTriggerRuntimeModule;
 pub use compat_get_default_export::CompatGetDefaultExportRuntimeModule;
 pub use create_fake_namespace_object::CreateFakeNamespaceObjectRuntimeModule;
 pub use create_script_url::CreateScriptUrlRuntimeModule;
-pub use css_loading::CssLoadingRuntimeModule;
 pub use define_property_getters::DefinePropertyGettersRuntimeModule;
 pub use ensure_chunk::EnsureChunkRuntimeModule;
 pub use export_webpack_require::ExportWebpackRequireRuntimeModule;
