+++ crates/rspack_testing/src/test_config.rs
@@ -503,7 +503,6 @@ impl TestConfig {
     plugins.push(rspack_plugin_asset::AssetPlugin.boxed());
     plugins.push(rspack_plugin_json::JsonPlugin {}.boxed());
     plugins.push(rspack_plugin_runtime::ArrayPushCallbackChunkFormatPlugin {}.boxed());
-    plugins.push(rspack_plugin_runtime::CssModulesPlugin {}.boxed());
     plugins.push(rspack_plugin_runtime::JsonpChunkLoadingPlugin {}.boxed());
     plugins.push(rspack_plugin_runtime::RuntimePlugin {}.boxed());
     if options.dev_server.hot {
